workon: Shell is now started using an exec instead of a subprocess.check_call.
For one thing, this is a good idea anyway, since there is no reason to keep running python after the shell has started.

For another thing, there are some shells like fish that somewhat expect this. If you don't start fish with an exec, it can't receive ^Cs, and if you type one, the whole shell just exits. So, that's not good.